Update README.md
Sentiment analysis, also refers as opinion mining, is a sub machine learning task where
we want to determine which is the general sentiment of a given document. Using machine
learning techniques and natural language processing we can extract the subjective information
of a document and try to classify it according to its polarity such as positive, neutral or negative.
It is a really useful analysis since we could possibly determine the overall opinion about a selling
objects, or predict stock markets for a given company like, if most people think positive about it,
possibly its stock markets will increase, and so on. Sentiment analysis is actually far from to be
solved since the language is very complex (objectivity/subjectivity, negation, vocabulary,
grammar,...) but it is also why it is very interesting to working on.
In this project I choose to try to classify tweets from Twitter into “positive” or “negative”
sentiment by building a model based on probabilities. Twitter is a microblogging website where
people can share their feelings quickly and spontaneously by sending a tweets limited by 140
characters. You can directly address a tweet to someone by adding the target sign “@” or
participate to a topic by adding an hastag “#” to your tweet. Because of the usage of Twitter, it is
a perfect source of data to determine the current overall opinion about anything.